+++ StateMachines/Runtime/Core/EventProviders/ITransitionEvaluatorEventsProvider.cs
@@ -1,0 +1,16 @@
+using System;
+
+namespace EggCentric.StateMachines
+{
+    public interface ITransitionEvaluatorEventsProvider
+    {
+        public event Action OnUninitializedStateMachine;
+        public event Action<Type> OnMissingRegisteredState;
+        public event Action<Type> OnRegisterStateDuplication;
+
+        public event Action<ITransition> OnTransitionAdded;
+
+        public event Action<Type, Type> OnMissingTransition;
+        public event Action<Type, Type> OnInvalidStateType;
+    }
+}