+++ StateMachines/Runtime/Core/ExecutionPolicies/IExecutionPolicy.cs
@@ -1,0 +1,9 @@
+namespace EggCentric.StateMachines
+{
+    public interface IExecutionPolicy
+    {
+        public TransitionFlags TransitionFlags { get; }
+
+        public abstract bool IsValid { get; }
+    }
+}