+++ StateMachines/Runtime/Core/Interfaces/States/IPlainState.cs
@@ -1,0 +1,7 @@
+namespace EggCentric.StateMachines
+{
+    public interface IPlainState : IState
+    {
+        public void Enter();
+    }
+}