+++ StateMachines/Runtime/States/Implementations/ICommonState.cs
@@ -1,7 +1,0 @@
-namespace EggCentric.StateMachines
-{
-    public interface ICommonState : IState
-    {
-        public void Enter();
-    }
-}