+++ StateMachines/CHANGELOG.md
@@ -1,3 +1,23 @@
+## [2.2.0] - 2025-11-28
+
+### Debug events and much more!
+
+- Implemented ability to get current state machine lock level
+- Implemented support of demug events for request receiving
+- Revamped debug events return types
+
+
+## [2.1.0] - 2025-11-28
+
+### Debug events and much more!
+
+- Implemented debugging events support
+- Implemented additional generic restriction for TransitionRequests
+- Implemented transition request ExecutionPolicies
+- Moved request execution and resolving logic into TransitionRequest 
+- Renamed ICommonState to IPlainState
+- Fixed transition request priority
+- Other improvements
 
 ## [2.0.1] - 2025-11-26